requirements.txt
Django
djangorestframework

django-model-utils

channels
channels_redis

psycopg2

flake8